Rubic binance bridge hot fixes (#5)
* add binance_bridge

* 111

* Added support for Celery task and some optimizations.

* Added few fixes.

* Updated panama_bridge views.

* ref cookie to query_params

* lower string all walletaddress

* check that actualFromAmount, actualToAmount returns by response from binance api

* update type of response time

* try to change output format to status and time

* Fixed transactions status.

* ref

* ref

* Delete docker-compose.yml

* Delete Dockerfile

* Minor fixes.

* Fix status codes.

* Fixed an amounts calculating.

* Added panama bridge admin.

* Fix checking transaction status.

* Minor fixes.

* ref image link

* refactor amount type

* Minor fixes.

* Minor fixes in decimals.

* changed preprocessing of fields amount

+++ lastwill/panama_bridge/models.py
@@ -10,8 +10,14 @@ class PanamaTransaction(models.Model):
 
     fromNetwork = models.CharField(max_length=4)
     toNetwork = models.CharField(max_length=4)
-    actualFromAmount = models.FloatField()
-    actualToAmount = models.FloatField()
+    actualFromAmount = models.DecimalField(
+        max_digits=50,
+        decimal_places=32
+    )
+    actualToAmount = models.DecimalField(
+        max_digits=50,
+        decimal_places=32
+    )
     ethSymbol = models.CharField(max_length=6)
     bscSymbol = models.CharField(max_length=6)
     updateTime = models.DateTimeField()
